+++ packages/core/textfield/src/components/__tests__/unit/server-side-hydrate.js
@@ -36,7 +36,6 @@ test('should ssr then hydrate textfield correctly', async () => {
           s === 'style'
         ),
     );
-
-    expect(mockCalls.length).toBe(0); // eslint-disable-line no-console
+    expect(mockCalls.length).toBe(0);
   });
 });